fix various problems

+++ bld/plusplus/c/fnname.c
@@ -378,7 +378,7 @@ static void appendBasedMod(     // APPEND A BASED MODIFIER
         } else {
             char    sbuf[60];
 
-            sprintf( sbuf, "internal name length=%d is > %d (1)", str->len, ZZ_LEN_LIMIT );
+            sprintf( sbuf, "internal name length=%d is > %d (1)", (int)str->len, ZZ_LEN_LIMIT );
             CFatal( sbuf );
         }
         appendStrWithLen( str->string, str->len );
